add delete materil feature

+++ libs/users/materials/feature-folder-list/src/lib/folders-list/folders-list.component.ts
@@ -4,6 +4,10 @@ import {FoldersEntity} from '@users/materials/data-access'
 import {FoldersFacade} from '@users/materials/data-access';
 import { Router } from '@angular/router';
 import { MatIconModule } from '@angular/material/icon';
+import { MatDialog } from '@angular/material/dialog';
+import {
+  FoldersDeleteDialogComponent
+} from '../folders-delete-dialog/folders-delete-dialog.component';
 
 @Component({
   selector: 'users-folders-state-list',
@@ -18,10 +22,22 @@ export class FoldersListComponent {
 
   private readonly foldersFacade: FoldersFacade = inject(FoldersFacade);
   private readonly router = inject(Router);
+  private readonly dialog: MatDialog = inject(MatDialog);
+
+  openDialog(id:number){
+    const dialogRef = this.dialog.open(FoldersDeleteDialogComponent)
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.foldersFacade.deleteFolder(id)
+      }
+    })
+
+  }
 
   deleteFolder(id: number, event: MouseEvent) {
     event.stopPropagation();
-    this.foldersFacade.deleteFolder(id);
+    this.openDialog(id)
+    // this.foldersFacade.deleteFolder(id);
   }
   openFolder() {
     this.router.navigate(['materials', this.folder.id]);
